email now functional and split up the code for future growth
